Remove CodecUtils
CodecUtils provides helper methods for serializing into a stream
compressed with zlib. The zlib compression is extreamly inefficient
and should generally be avoided in favor of ZSTD.

To avoid dependency bloat it is better to avoid adding zstd as a dependency
for the airlift-json module, thus it was decided to simply remove these
utility methods.