Create a generic thrift response for http response with error code.
Currently thrift response handler is leveraging thift codec for all responses except with http status = 4xx (user error). So for cases where there are server errors(500) or timeout issues(503) etc, the response handler is broken as the codec does not understand the entity embedded into the http response. The change as part of this commit is to address this issue.